# openapi.yaml
openapi: 3.1.0
info:
  title: Rolay Server API
  version: 0.3.0
  description: >
    HTTP contract for Rolay v1. The product model is room-based:
    users have global roles, rooms have owner/member roles, and room trees
    sync through snapshot + SSE + CRDT document sessions. Excalidraw drawings
    use blob-backed persistence plus single-editor live sessions.
servers:
  - url: https://rolay.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/rooms:
    get:
      tags: [rooms]
      summary: List rooms for current user
      operationId: listRooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User room list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
    post:
      tags: [rooms]
      summary: Create room
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /v1/rooms/{workspaceId}/members:
    get:
      tags: [rooms]
      summary: List room members for the current participant
      operationId: listRoomMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Room members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomMemberListResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /v1/rooms/{workspaceId}/publication:
    get:
      tags: [rooms]
      summary: Get room publication state
      operationId: getRoomPublication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Room publication state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPublicationResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    patch:
      tags: [rooms]
      summary: Toggle room publication state
      operationId: updateRoomPublication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomPublicationRequest'
      responses:
        '200':
          description: Updated room publication state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPublicationResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /v1/workspaces:
    get:
      tags: [rooms]
      summary: List rooms for current user (workspace alias)
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User room list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
    post:
      tags: [rooms]
      summary: Create room (workspace alias)
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /v1/workspaces/{workspaceId}/members:
    get:
      tags: [rooms]
      summary: List room members for the current participant (workspace alias)
      operationId: listWorkspaceMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Room members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomMemberListResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /v1/workspaces/{workspaceId}/publication:
    get:
      tags: [rooms]
      summary: Get room publication state (workspace alias)
      operationId: getWorkspacePublication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Room publication state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPublicationResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    patch:
      tags: [rooms]
      summary: Toggle room publication state (workspace alias)
      operationId: updateWorkspacePublication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomPublicationRequest'
      responses:
        '200':
          description: Updated room publication state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPublicationResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /public/api/rooms/{workspaceId}/manifest:
    get:
      tags: [public]
      summary: Get public room manifest
      operationId: getPublicRoomManifest
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Public room manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicRoomManifest'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /public/api/rooms/{workspaceId}/events:
    get:
      tags: [public]
      summary: Subscribe to public room events
      operationId: streamPublicRoomEvents
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Public server-sent event stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  id: 42
                  event: tree.entry.updated
                  data: {"entryId":"fil_1","path":"Week-01.md","entryVersion":3}
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /public/api/rooms/{workspaceId}/markdown/{entryId}/crdt-token:
    post:
      tags: [public]
      summary: Create a public read-only Markdown CRDT token
      operationId: createPublicMarkdownCrdtToken
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/EntryId'
      responses:
        '200':
          description: Read-only public CRDT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicCrdtTokenResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /public/api/rooms/{workspaceId}/files/{entryId}/blob/content:
    get:
      tags: [public]
      summary: Read a public image or Excalidraw blob
      operationId: getPublicBlobContent
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/EntryId'
        - name: hash
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full blob content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial blob content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
      description: Room identifier.
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
      description: Stable file entry identifier.
...
  responses:
    ErrorBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorNotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    GlobalRole:
      type: string
      enum: [admin, writer, reader]
...
    WorkspaceRole:
      type: string
      enum: [owner, member]
    User:
      type: object
      additionalProperties: false
      required: [id, username, displayName, isAdmin, globalRole]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        globalRole:
          $ref: '#/components/schemas/GlobalRole'
...
    Workspace:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
    WorkspacePublication:
      type: object
      additionalProperties: false
      required: [enabled, updatedAt]
      properties:
        enabled:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    RoomPublication:
      type: object
      additionalProperties: false
      required: [workspaceId, enabled, updatedAt]
      properties:
        workspaceId:
          type: string
        enabled:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    RoomPublicationResponse:
      type: object
      additionalProperties: false
      required: [publication]
      properties:
        publication:
          $ref: '#/components/schemas/RoomPublication'
    UpdateRoomPublicationRequest:
      type: object
      additionalProperties: false
      required: [enabled]
      properties:
        enabled:
          type: boolean
    WorkspaceResponse:
      type: object
      additionalProperties: false
      required: [workspace]
      properties:
        workspace:
          $ref: '#/components/schemas/Workspace'
    CreateRoomRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          deprecated: true
    RoomListItem:
      type: object
      additionalProperties: false
      required: [workspace, membershipRole, createdAt, memberCount, inviteEnabled, publication]
      properties:
        workspace:
          $ref: '#/components/schemas/Workspace'
        membershipRole:
          $ref: '#/components/schemas/WorkspaceRole'
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
          minimum: 0
        inviteEnabled:
          type: boolean
        publication:
          $ref: '#/components/schemas/WorkspacePublication'
...
    RoomListResponse:
      type: object
      additionalProperties: false
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/RoomListItem'
...
    RoomMember:
      type: object
      additionalProperties: false
      required: [user, role, joinedAt]
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        joinedAt:
          type: string
          format: date-time
    RoomMemberListResponse:
      type: object
      additionalProperties: false
      required: [members]
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/RoomMember'
...
    BlobRevision:
      type: object
      additionalProperties: false
      required: [hash, sizeBytes, mimeType]
      properties:
        hash:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        mimeType:
          type: string
...
    PublicManifestEntry:
      type: object
      additionalProperties: false
      required: [id, path, kind, contentMode, entryVersion, updatedAt]
      properties:
        id:
          type: string
        path:
          type: string
        kind:
          type: string
          enum: [folder, markdown, excalidraw]
        contentMode:
          type: string
          enum: [none, crdt, blob]
        entryVersion:
          type: integer
          minimum: 0
        docId:
          type: string
        blob:
          $ref: '#/components/schemas/BlobRevision'
        updatedAt:
          type: string
          format: date-time
    PublicAsset:
      type: object
      additionalProperties: false
      required: [entryId, path, hash, sizeBytes, mimeType, contentUrl]
      properties:
        entryId:
          type: string
        path:
          type: string
        hash:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        mimeType:
          type: string
        contentUrl:
          type: string
    PublicRoomManifest:
      type: object
      additionalProperties: false
      required: [workspace, publication, cursor, entries, assets]
      properties:
        workspace:
          $ref: '#/components/schemas/Workspace'
        publication:
          $ref: '#/components/schemas/WorkspacePublication'
        cursor:
          type: integer
          minimum: 0
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PublicManifestEntry'
        assets:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/PublicAsset'
    PublicCrdtTokenResponse:
      allOf:
        - $ref: '#/components/schemas/CrdtTokenResponse'
        - type: object
          additionalProperties: false
          required: [readOnly]
          properties:
            readOnly:
              type: boolean
              const: true
...
    CrdtTokenResponse:
      type: object
      additionalProperties: false
      required: [entryId, docId, provider, wsUrl, token, expiresAt]
      properties:
        entryId:
          type: string
        docId:
          type: string
        provider:
          type: string
          example: yjs-hocuspocus
        wsUrl:
          type: string
          format: uri
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - unauthorized
            - forbidden
            - workspace_not_found
            - invite_not_found
            - invite_disabled
            - user_not_found
            - entry_not_found
            - path_already_exists
            - entry_version_mismatch
            - username_taken
            - invalid_role
            - cannot_delete_admin_user
        message:
          type: string
        details:
          type: object
          additionalProperties: true